feat: 모니터링 설정 키워드, 이메일 페이지 분리
